tensors fixes

+++ data/dataset.py
@@ -149,8 +149,8 @@ class CsvDataset(object):
         y1 = target_csv.iloc[item,2] * img_size[1]
         x2 = target_csv.iloc[item, 3] * img_size[0]
         y2 = target_csv.iloc[item, 4] * img_size[1]
-        box = torch.LongTensor(np.array([x1, y1, x2, y2], dtype=np.long))
-        return img, box, label, torch.LongTensor(1)
+        box = torch.FloatTensor(np.array([x1, y1, x2, y2], dtype=np.float))
+        return img, box, label, torch.FloatTensor(1)
 
 
 class Dataset:
